CACHE-13371: fix request cf cache control tests
